5767 - Split hooks: useGetItems and useItems
