+++ src/client/pages/CountryHome/Repository/RepositoryList/RepositoryList.tsx
@@ -1,15 +1,19 @@
 import React from 'react'
 
-import { useItems } from 'client/pages/CountryHome/Repository/RepositoryList/hooks/useItems'
 import RepositoryListItem from 'client/pages/CountryHome/Repository/RepositoryList/RepositoryListItem'
 
+import { useGetItems } from './hooks/useGetItems'
+import { useItems } from './hooks/useItems'
 import { useOnToggle } from './hooks/useOnToggle'
+
 type Props = {
   isGlobal?: boolean
 }
 
 const RepositoryList: React.FC<Props> = (props) => {
   const { isGlobal } = props
+
+  useGetItems(isGlobal)
   const items = useItems(isGlobal)
   const { collapsed, onToggle } = useOnToggle()
